fix natural_t to be uint32_t
(cherry picked from commit d775a5a7b1588ae04426362f0f3ee6e65aea685c)

+++ include/mach/i386/vm_types.h
@@ -100,7 +100,7 @@
  * a port in user space as an integer and
  * in kernel space as a pointer.
  */
-typedef unsigned long	natural_t;
+typedef uint32_t	natural_t;
 
 /*
  * An integer_t is the signed counterpart
